Stake table contract and staking CLI (#2662)
* New stake table contract skeleton

This is not a working implementation but intended to provide rust
bindings that are equal or close to what we expect after implementation
is done.

* new stake table storage layout

This is draft and there are no tests yet.

* fix typo

* remove redundant zero check for BLS key

We are already checking a signature from the BLS key. Due to pre-image
resistance we don't need to check if the key is zero.

* remove unused function

* ensure the same BLS key cannot be registered twice

* fixup contract bindings after merging in main

* add back light client to stake table contract

* re-enable some stake table tests

- Some of the tests do not apply anymore since the requirements changed.
  Notably anything relating to the queue logic for entering and exiting.
- Fix that we were not crediting delegations.
- Update undelegation test to test partial withdrawals.

* Draft: make stake table contract upgradable

This is not yet tested but I would like to add the rust code to deploy
the contract and for that I need the contract to be upgradable so we can
deploy it behind a proxy.

* Make stake table contract upgradable

- Make token contract upgradable.
- Use proxy in stake table tests.
- Add forge code to deploy stake table and token.
- Add rust code to deploy token (stake table still WIP).

* remove stake table ethers bindings

We don't want to use ethers-rs for new code anymore.

* update forge-std

* feat: rust code to deploy stake table contract

Keeps the old permissioned stake table around for now so we don't break
everything that relies on that. We can remove it later.

Update some contract addresses since the number of contracts deployed
has changed.

Add argument to set the contract exit escrow period to the deployer CLI.

* Add partial implementation for staking CLI

* remove commented code

* more test cases for parse commission

* Add test system for staking-cli

* staking-cli: register a validator

Includes BLS signature on address. The remaining functions should be
rather easy in comparison.

* add TODO test: undelegation after validator exit

* commit missing files

* test: validator registration via CLI

* staking-cli: implement delegation

* staking-cli: implement undelegation

* stake table contract: index events

* staking-cli: implement withdrawals

* staking-cli: add command to stake for the demo

* Remove the abstract stake table contract

The contract is an unnecessary indirection and doesn't implement
anything.

Add natspec docs to stake table contract.

* Fix compilation of dev node

- Remove some unused imports.

* fix compilation

* Make CLI tests rust integration tests

* mention why `info` command is not yet implemented

* fix: clippy

* remove commented code

* forge fmt

* Update staking-cli/README.md

Co-authored-by: tbro <48967308+tbro@users.noreply.github.com>

* revert on zero addresses during initialization

* remove unused setupCustom functon

* fix: lookup validator, not delegator

Update integration test to ensure it fails in this scenario.

* More efficient storage layout

Put small enum after the full word uint256

Update bindings, they were out of date.

* Update staking-cli/README.md

Co-authored-by: tbro <48967308+tbro@users.noreply.github.com>

---------

Co-authored-by: Abdul Basit <45506001+imabdulbasit@users.noreply.github.com>
Co-authored-by: tbro <48967308+tbro@users.noreply.github.com>