Export a proper library from `staking-cli` (#2857)
Export a proper library from `staking-cli`

`staking-cli` has a lot of functions we can use in `TestNetwork`. In
order to use them we need it to act as a library. This PR also adds
the lib to `sequencer` crate for use in follow up PRs.

Note that we should consider moving this logic into `contracts/adapter` crate in the future, and importing everything from there.

---------

Co-authored-by: tbro <tbro@users.noreply.github.com>